backend/requirements.txt
Django>=5.0.3
djangorestframework>=3.14.0
django-environ>=0.11.2
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0
django-widget-tweaks>=1.5.0
Pillow>=10.3.0
requests>=2.31.0
reportlab>=4.2.2
